Hook the LFGList update function to generate buttons
This allows to remove some code, as well as stop relying on events, meaning that all button updates should now be instant instead of depending on the event to happen. This also allows not to override the default ScrollFrame update function, which should solve the addon taking the costs of blizz's stuff on itself.

Yay, hooks!