Do not nil out buttons
Frames cannot be deleted, nil-ing a `LFClean.buttons` entry just removes the reference but the frame still sticks around. This caused frames to just getting generated anew but never actually getting deleted when the option was toggled, also causing the issue of the icon getting blacked out after toggling the entry buttons option a couple of times with the group finder open. Instead of nil-ing the buttons, just hiding them solves the issue.